+++ structurizr-core/src/main/java/com/structurizr/view/CustomView.java
@@ -110,7 +110,7 @@ public final class CustomView extends ModelView implements AnimatedView {
             }
         }
 
-        if (elementsInThisAnimationStep.size() == 0) {
+        if (elementsInThisAnimationStep.isEmpty()) {
             throw new IllegalArgumentException("None of the specified elements exist in this view.");
         }