+++ structurizr-dsl/src/test/java/com/structurizr/dsl/CustomViewAnimationStepParserTests.java
@@ -1,5 +1,6 @@
 package com.structurizr.dsl;
 
+import com.structurizr.view.CustomView;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -15,7 +16,7 @@ class CustomViewAnimationStepParserTests extends AbstractTests {
             parser.parse((CustomViewDslContext)null, tokens("animationStep"));
             fail();
         } catch (Exception e) {
-            assertEquals("Expected: animationStep <identifier> [identifier...]", e.getMessage());
+            assertEquals("Expected: animationStep <identifier|element expression> [identifier|element expression...]", e.getMessage());
         }
     }
 
@@ -25,7 +26,23 @@ class CustomViewAnimationStepParserTests extends AbstractTests {
             parser.parse((CustomViewAnimationDslContext) null, tokens());
             fail();
         } catch (Exception e) {
-            assertEquals("Expected: <identifier> [identifier...]", e.getMessage());
+            assertEquals("Expected: <identifier|element expression> [identifier|element expression...]", e.getMessage());
+        }
+    }
+
+    @Test
+    void test_parse_ThrowsAnException_WhenTheElementDoesNotExist() {
+        CustomView view = workspace.getViews().createCustomView("key", "Title", "Description");
+
+        CustomViewAnimationDslContext context = new CustomViewAnimationDslContext(view);
+        IdentifiersRegister map = new IdentifiersRegister();
+        context.setIdentifierRegister(map);
+
+        try {
+            parser.parse(context, tokens("e"));
+            fail();
+        } catch (Exception e) {
+            assertEquals("The element/relationship \"e\" does not exist", e.getMessage());
         }
     }