+++ structurizr-dsl/src/test/java/com/structurizr/dsl/StaticViewAnimationStepParserTests.java
@@ -1,5 +1,6 @@
 package com.structurizr.dsl;
 
+import com.structurizr.view.SystemLandscapeView;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -15,7 +16,7 @@ class StaticViewAnimationStepParserTests extends AbstractTests {
             parser.parse((StaticViewDslContext)null, tokens("animationStep"));
             fail();
         } catch (Exception e) {
-            assertEquals("Expected: animationStep <identifier> [identifier...]", e.getMessage());
+            assertEquals("Expected: animationStep <identifier|element expression> [identifier|element expression...]", e.getMessage());
         }
     }
 
@@ -25,7 +26,23 @@ class StaticViewAnimationStepParserTests extends AbstractTests {
             parser.parse((StaticViewAnimationDslContext) null, tokens());
             fail();
         } catch (Exception e) {
-            assertEquals("Expected: <identifier> [identifier...]", e.getMessage());
+            assertEquals("Expected: <identifier|element expression> [identifier|element expression...]", e.getMessage());
+        }
+    }
+
+    @Test
+    void test_parse_ThrowsAnException_WhenTheElementDoesNotExist() {
+        SystemLandscapeView view = workspace.getViews().createSystemLandscapeView("key", "Description");
+
+        StaticViewAnimationDslContext context = new StaticViewAnimationDslContext(view);
+        IdentifiersRegister map = new IdentifiersRegister();
+        context.setIdentifierRegister(map);
+
+        try {
+            parser.parse(context, tokens("user"));
+            fail();
+        } catch (Exception e) {
+            assertEquals("The element/relationship \"user\" does not exist", e.getMessage());
         }
     }